soc: imx: blk-ctl: Add hack to keep DSI clocks running
It looks like the DSIM driver does not enable the clocks properly which
causes the boot to lockup at probe time, as the BLK CTL is disabling the
clocks after usage (which is probably correct). For now add this hack to
keep the clocks running until we have a proper solution in the DSIM driver.

Signed-off-by: Frieder Schrempf <frieder.schrempf@kontron.de>